Add styling to student-details
